@@ -1,13 +1,15 @@
 {
-  "name": "wstrade-api",
+  "name": "wstrade-api-customized",
   "version": "1.10.2",
   "description": "Wealthsimple Trade API Wrapper",
   "main": "dist.js",
   "types": "index.d.ts",
+  "type": "module",
   "scripts": {
     "bundle": "webpack",
     "lint": "npx eslint --fix src/ --ignore-pattern __test__",
-    "test": "jest"
+    "test": "jest",
+    "start": "node src/app.js"
   },
   "repository": {
     "type": "git",
@@ -34,18 +36,20 @@
     "@babel/preset-env": "^7.9.6",
     "babel-jest": "^26.6.3",
     "babel-loader": "^8.2.2",
-    "eslint": "^7.17.0",
+    "eslint": "^7.32.0",
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-react": "^7.34.2",
     "jest": "^26.6.3",
     "typescript": "^3.9.2",
-    "webpack": "^5.4.0",
+    "webpack": "^5.92.0",
     "webpack-cli": "^4.2.0"
   },
   "dependencies": {
     "cloudscraper": "^4.6.0",
-    "node-fetch": "^2.6.1",
+    "node-fetch": "^2.7.0",
     "request": "^2.88.2",
-    "source-map-support": "^0.5.19"
+    "source-map-support": "^0.5.19",
+    "wstrade-api": "^1.10.2"
   }
 }
